# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SpotMoney Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserInit {
  telegramId: string;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  referredBy?: string | null;
}

export interface User {
  id: number;
  telegramId: string;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  balance: number;
  totalEarned: number;
  adsWatchedToday: number;
  /** @nullable */
  adCooldownUntil?: string | null;
  streak: number;
  referralCode: string;
  createdAt: string;
}

export interface AdRewardResult {
  balance: number;
  coinsEarned: number;
  adsWatchedToday: number;
  cooldownUntil: string;
}

export interface ReferralUser {
  firstName: string;
  /** @nullable */
  username?: string | null;
  joinedAt: string;
}

export interface ReferralInfo {
  referralCode: string;
  referralLink: string;
  totalReferrals: number;
  referralEarnings: number;
  referrals: ReferralUser[];
}

export interface UserStats {
  balance: number;
  totalEarned: number;
  adsWatchedTotal: number;
  streak: number;
  leaderboardRank: number;
  referralCount: number;
}

export interface Task {
  id: number;
  title: string;
  description: string;
  reward: number;
  icon: string;
  actionUrl: string;
  category: string;
}

export interface TaskWithStatus {
  id: number;
  title: string;
  description: string;
  reward: number;
  icon: string;
  actionUrl: string;
  category: string;
  completed: boolean;
}

export interface TaskResult {
  balance: number;
  coinsEarned: number;
  taskId: number;
}

export interface WithdrawalInput {
  amount: number;
  method: string;
  address: string;
}

export interface Withdrawal {
  id: number;
  amount: number;
  method: string;
  address: string;
  status: string;
  createdAt: string;
}

export interface LeaderboardEntry {
  rank: number;
  firstName: string;
  /** @nullable */
  username?: string | null;
  totalEarned: number;
  streak: number;
}