# lib/api-zod/src/generated/types/adRewardResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SpotMoney Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */

export interface AdRewardResult {
  balance: number;
  coinsEarned: number;
  adsWatchedToday: number;
  cooldownUntil: string;
}