# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SpotMoney Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  rank: number;
  firstName: string;
  /** @nullable */
  username?: string | null;
  totalEarned: number;
  streak: number;
}