# lib/api-zod/src/generated/types/referralInfo.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SpotMoney Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */
import type { ReferralUser } from './referralUser';

export interface ReferralInfo {
  referralCode: string;
  referralLink: string;
  totalReferrals: number;
  referralEarnings: number;
  referrals: ReferralUser[];
}