# lib/api-zod/src/generated/types/referralUser.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SpotMoney Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */

export interface ReferralUser {
  firstName: string;
  /** @nullable */
  username?: string | null;
  joinedAt: string;
}