# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SpotMoney Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */

export interface Task {
  id: number;
  title: string;
  description: string;
  reward: number;
  icon: string;
  actionUrl: string;
  category: string;
}