# lib/api-zod/src/generated/types/taskResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SpotMoney Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskResult {
  balance: number;
  coinsEarned: number;
  taskId: number;
}