# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SpotMoney Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  telegramId: string;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  balance: number;
  totalEarned: number;
  adsWatchedToday: number;
  /** @nullable */
  adCooldownUntil?: string | null;
  streak: number;
  referralCode: string;
  createdAt: string;
}