# lib/api-zod/src/generated/types/userInit.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SpotMoney Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */

export interface UserInit {
  telegramId: string;
  firstName: string;
  /** @nullable */
  lastName?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  referredBy?: string | null;
}