# lib/api-zod/src/generated/types/userStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SpotMoney Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */

export interface UserStats {
  balance: number;
  totalEarned: number;
  adsWatchedTotal: number;
  streak: number;
  leaderboardRank: number;
  referralCount: number;
}