# lib/api-zod/src/generated/types/withdrawal.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * SpotMoney Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */

export interface Withdrawal {
  id: number;
  amount: number;
  method: string;
  address: string;
  status: string;
  createdAt: string;
}